+++ packages/runtime/adapters/eventstore-adapters/eventstore-base/src/cursor-operations.ts
@@ -4,7 +4,15 @@ import {
   CURSOR_BUFFER_SIZE,
   THREAD_COUNTER_BYTE_LENGTH,
 } from './constants'
-import { Cursor, EventsWithCursor } from './types'
+import { Cursor, EventsWithCursor, EventWithCursor } from './types'
+
+const checkThreadArrayLength = (threadArray: Array<number>): void => {
+  assert.strictEqual(
+    threadArray.length,
+    THREAD_COUNT,
+    'Cursor must be represented by array of 256 numbers'
+  )
+}
 
 const checkThreadCounterHexString = (threadCounter: string): void => {
   assert.strictEqual(
@@ -21,11 +29,7 @@ export const initThreadArray = (): Array<number> => {
 }
 
 export const threadArrayToCursor = (threadArray: Array<number>): string => {
-  assert.strictEqual(
-    threadArray.length,
-    THREAD_COUNT,
-    'Cursor must be represented by array of 256 numbers'
-  )
+  checkThreadArrayLength(threadArray)
 
   const cursorBuffer: Buffer = Buffer.alloc(CURSOR_BUFFER_SIZE)
 
@@ -97,4 +101,60 @@ export const emptyLoadEventsResult = (cursor: Cursor): EventsWithCursor => {
     cursor: cursor == null ? threadArrayToCursor(initThreadArray()) : cursor,
     events: [],
   }
+}
+
+const calculateMaxThreadArray = (
+  threadArrays: Array<Array<number>>
+): Array<number> => {
+  const maxThreadArray = initThreadArray()
+  for (const threadArray of threadArrays) {
+    checkThreadArrayLength(threadArray)
+    for (let i = 0; i < THREAD_COUNT; ++i) {
+      maxThreadArray[i] = Math.max(maxThreadArray[i], threadArray[i])
+    }
+  }
+  return maxThreadArray
+}
+
+export const checkEventsContinuity = (
+  startingCursor: Cursor,
+  eventCursorPairs: EventWithCursor[]
+): boolean => {
+  const startingThreadArray = cursorToThreadArray(startingCursor)
+
+  const tuples = eventCursorPairs.map(({ event, cursor }) => {
+    return {
+      event,
+      cursor,
+      threadArray: cursorToThreadArray(cursor),
+    }
+  })
+
+  for (let i = 0; i < tuples.length; ++i) {
+    assert.strictEqual(
+      tuples[i].event.threadCounter,
+      tuples[i].threadArray[tuples[i].event.threadId] - 1
+    )
+    if (
+      startingThreadArray[tuples[i].event.threadId] >
+      tuples[i].event.threadCounter
+    ) {
+      return false
+    }
+  }
+
+  const maxThreadArray = calculateMaxThreadArray(
+    tuples.map((t) => t.threadArray)
+  )
+
+  for (const t of tuples) {
+    startingThreadArray[t.event.threadId]++
+  }
+
+  for (let i = 0; i < THREAD_COUNT; ++i) {
+    if (maxThreadArray[i] !== startingThreadArray[i]) {
+      return false
+    }
+  }
+  return true
 }